requirements.txt
asgiref==3.6.0
attrs==23.1.0
Brotli==1.0.9
colorama==0.4.6
dj-database-url==2.0.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
gunicorn==20.1.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
packaging==23.1
pluggy==1.0.0
psycopg==3.1.9
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.3.1
pytest-django==4.5.2
pytest-testdox==3.0.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
whitenoise==6.4.0
